all crud methods added

+++ src/main/java/kz/kbtu/phonebook/repository/UserRepository.java
@@ -4,7 +4,5 @@ import org.springframework.data.domain.Pageable;
 import kz.kbtu.phonebook.model.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
-//    void create(User user);
-//    User getByUsername(String username);
+public interface UserRepository extends CrudRepository<User, Long> {
 }
